terraform/modules/backup/requirements.txt
google-cloud-sql==3.4.4
google-cloud-storage==2.10.0
google-cloud-pubsub==2.18.4
google-cloud-monitoring==2.16.0